Restructured Loading Tests
