+++ applications/StructuralMechanicsApplication/tests/test_loading_conditions_point.py
@@ -1,0 +1,16 @@
+from __future__ import print_function, absolute_import, division
+import KratosMultiphysics 
+
+import KratosMultiphysics.StructuralMechanicsApplication as StructuralMechanicsApplication
+import KratosMultiphysics.KratosUnittest as KratosUnittest
+import math
+
+
+class TestLoadingConditionsPoint(KratosUnittest.TestCase):
+    def setUp(self):
+        pass
+    
+    
+        
+if __name__ == '__main__':
+    KratosUnittest.main()
